Add files via upload
Strengths of Media Queries: Extends content to a broader viewing audience of a variety of hardware formats.
Weakness: If not actively updated to new tech standards, it can turn against you as your site loses visual fidelity on newer devices.

Observations: Sculpting a site is not too hard but takes a while. Learning Git Hub does not take long, but is really freak'n hard.